requirements.txt
Django==2.2.6
django-bootstrap4==1.0.1
django-filter==2.2.0
django-heroku==0.3.1
django-impersonate==1.4.1
django-mathfilters==0.4.0
django-tables2==2.1.1
gunicorn==19.9.0
social-auth-app-django==3.1.0
whitenoise==4.1.4
beautifulsoup4==4.8.1
html5lib==1.0.1